LayoutTests/fast/js/console-non-string-values.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<link rel="stylesheet" href="resources/js-test-style.css">
<script src="resources/js-test-pre.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>
<script>
debug('Should see two console messages, the first a string, and the second a number.');
window.console.log('I am a string');
window.console.log(42);
var successfullyParsed = true;
</script>
<script src="resources/js-test-post.js"></script>
</body>
</html>
