LayoutTests/fast/js/console-non-string-values-expected.txt

CONSOLE MESSAGE: line 0: I am a string
CONSOLE MESSAGE: line 0: 42
Should see two console messages, the first a string, and the second a number.
PASS successfullyParsed is true

TEST COMPLETE
